Synced sequelize with the database, making the associations work

+++ migrations/20220719003427-create-post.js
@@ -36,7 +36,7 @@ module.exports = {
         defaultValue: Date.now()
       },
       user_id: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.SMALLINT,
         allowNull: false
       }
     });
