+++ migrations/20220720005354-create-comment.js
@@ -33,11 +33,11 @@ module.exports = {
         defaultValue: Date.now()
       },
       user_id: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.SMALLINT,
         allowNull: false
       },
       post_id: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.SMALLINT,
         allowNull: false
       }
     });
